third_party/blink/web_tests/http/tests/websocket/client-close-expected.txt

WebSocket: Test client-initiated close. After WebSocket.close() any message from server must be discarded by WebSocket stack according to the new WebSocket API. See http://webkit.org/b/85934.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
Closed
PASS receivedMessage is undefined.
PASS closeEvent.reason is "close_frame[:2]='88 80'"
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
